add training/testing split and stantargets. Running locally.

+++ R/sync_envdata.R
@@ -1,8 +1,9 @@
 
 
 sync_envdata <- function(repo="AdamWilsonLab/emma_envdata",
-                         path="data/envdata"){
-  file.remove(list.files(path,full=T)) # delete all old files
+                         path="data/envdata",delete_all_first=F){
+if(delete_all_first)  file.remove(list.files(path,full=T)) # delete all old files
+
   #file.path(path,"stable_data.gz.parquet"))
   if(!file.exists(path)) dir.create(path,recursive = T)
   piggyback::pb_download(repo = repo,
